requirements.txt
streamlit>=1.40.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.0.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
python-pptx>=1.0.0
yt-dlp>=2024.0.0
openpyxl>=3.1.0
tenacity>=9.0.0
